+++ docs/CHANGELOG.md
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+2.1.0  (December 7, 2021)
+-------------------------
+- Enh #564: Restrict messaging for blocked users
+- Enh #232: Render images in email notifications
+- Fix #224: Fix preview message in sidebar
+- Fix #231: CLI error when no REST module is installed
+- Fix #230: Fix notifications of new Conversation vs Message entry
+- Fix #240: Fix apostrophe encode in mail
+
 2.0.7  (April 8, 2021)
 ----------------------
 - Fix #221: Fix call of console commands when REST API module doesn't exist